move bugs to separate package
