+++ compose/mpp/demo/src/commonMain/kotlin/androidx/compose/mpp/demo/bug/RoundedCornerCrashOnJS.kt
@@ -14,18 +14,17 @@
  * limitations under the License.
  */
 
-package androidx.compose.mpp.demo
+package androidx.compose.mpp.demo.bug
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.runtime.Composable
+import androidx.compose.mpp.demo.Screen
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.unit.dp
 
-@Composable
-fun RoundedCornerCrashOnJS() {
+val RoundedCornerCrashOnJS = Screen.Example("RoundedCornerCrashOnJS") {
     // Crash happens in ShadowUtils.drawShadow(
     // Related issue https://github.com/JetBrains/compose-multiplatform/issues/3013
     Box(
